Updated model binding to support fields as well as properties.
Added ModelBinding/BindingMemberInfo.cs to Nancy.csproj, migrating into it the logic DefaultBinder.cs uses to enumerate properties and expanding it to support fields as well. Added unit testing class Unit/ModelBinding/BindingMemberInfoFixture.cs to Nancy.Tests.csproj.
Updated ModelBinding/BindingContext.cs to store an enumerable sequence of BindingMemberInfo objects instead of directly referencing System.Reflection.PropertyInfo. Updated all unit tests referencing this field accordingly.
Updated ModelBinding/DefaultBinder.cs to work with BindingMemberInfo objects wherever it refers to model properties. The primary impact of this change is that .GetValue and .SetValue calls no longer need to supply a dummy 'null' parameter for the property's index arguments, as this detail (which does not apply, in any case, to fields) is now hidden within BindingMemberInfo.
Updated the initialization of BindingContext objects in ModelBinding/DefaultBinder.cs to convert the enumerable list of properties to a list before stashing it in the ValidModelProperties member, because it does get enumerated more than once in some circumstances.
Reworked the GetProperties helper method in ModelBinding/DefaultBinder.cs to leverage the new BindingMemberInfo.Collect method, leaving only the blacklist to be implemented within the method.
Updated the Unit/ModelBinding/DefaultBinderFixture.cs and Unit/ModelBinding/DefaultBodyDeserializers/JsonBodyDeserializerFixture.cs unit test classes in Nancy.Tests to verify the functionality of the BindingMemberInfo objects on both properties and fields.
Refactored the copy/paste overloads of BindProperty in ModelBinding/DefaultBinder.cs so that one overload calls the other.
Fixed a bug in the BindProperty overload in ModelBinding/DefaultBinder.cs that takes a specified instance for the target of the binding that may have caused certain assignments to go to the wrong object.
Renamed GetProperties to GetBindingMembers, BindProperty to BindValue and SetPropertyValue to SetBindingMemberValue in ModelBinding/DefaultBinder.cs.