Updated code to match style guidelines, and improved naming.
Refactored the copy/paste overloads of BindProperty in ModelBinding/DefaultBinder.cs so that one overload calls the other.
Fixed a bug in the BindProperty overload in ModelBinding/DefaultBinder.cs that takes a specified instance for the target of the binding that may have caused certain assignments to go to the wrong object.
Renamed GetProperties to GetBindingProperties, BindProperty to BindValue and SetPropertyValue to SetBindingPropertyValue in ModelBinding/DefaultBinder.cs.
Replaced tabs with spaces, moved 'using' statements inside the namespace, added braces around single-line 'if' bodies, removed underscores from members and added explicit 'this.' qualifier to member references in ModelBinding/BindingPropertyInfo.cs in Nancy.csproj and Unit/ModelBinding/BindingPropertyInfoFixture.cs in Nancy.Tests.csproj.